spring boot commit

+++ inventory-service/src/main/java/com/example/inventoryservice/Dto/DiscountDto.java
@@ -1,6 +1,6 @@
 package com.example.inventoryservice.Dto;
 
-import com.example.inventoryservice.Entities.Discount;
+
 import lombok.Builder;
 import lombok.Data;
 
@@ -9,6 +9,6 @@ import lombok.Data;
 @Builder
 public class DiscountDto {
     private Long id;
-    private double discountPercentage;
+    private int discountPercentage;
     private String productId;
 }
