+++ inventory-service/src/main/java/com/example/inventoryservice/Dto/ProductDto.java
@@ -12,5 +12,5 @@ public class ProductDto {
     private String description;
     private String quality;
     private int price;
-    private double percentage;
+    private int percentage;
 }
